+++ src/settings/MatchProfiles/MatchProfiles.js
@@ -27,6 +27,7 @@ import {
   ORDER_RESOURCE_PATHS,
   NOTES_RESOURCE_PATHS,
   INVOICE_RESOURCE_PATHS,
+  ACQ_DATA_RESOURCE_PATHS,
   FIND_ALL_CQL,
   OCLC_MATCH_EXISTING_SRS_RECORD_ID,
   OCLC_MATCH_NO_SRS_RECORD_ID,
@@ -386,9 +387,11 @@ export class MatchProfiles extends Component {
       const requestsToOrder = ORDER_RESOURCE_PATHS.map(path => fetchJsonSchema(path, ordersModuleVersion, okapi));
       const requestsToNotes = NOTES_RESOURCE_PATHS.map(path => fetchJsonSchema(path, notesModuleVersion, okapi));
       const requestsToInvoice = INVOICE_RESOURCE_PATHS.map(path => fetchJsonSchema(path, invoiceModuleVersion, okapi));
+      const requestToAcquisitionsData = ACQ_DATA_RESOURCE_PATHS.map(path => fetchJsonSchema(path, ordersModuleVersion, okapi));
 
       await handleAllRequests(requestsToInstance, 'INSTANCE', this.addToState);
       await handleAllRequests(requestsToHoldings, 'HOLDINGS', this.addToState);
+      await handleAllRequests(requestToAcquisitionsData, 'HOLDINGS', this.addToState);
       await handleAllRequests(requestsToItem, 'ITEM', this.addToState);
       await handleAllRequests(requestsToOrder, 'ORDER', this.addToState);
       await handleAllRequests(requestsToNotes, 'ORDER', this.addToState);